Tighten requirements around unique names
To make names easier to share, copy and paste in all contexts, we
limit them to alphanumeric characters and hyphens.

This allows names to be easily and unambiguously parsed by applications.

The alternative would have been to disallow certain characters in
certain contexts, which is much more error prone and it would be easy
to forget about certain edge cases.